requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.1.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
dj-database-url==0.4.2
Django==2.0.7
django-celery-results==1.0.1
django-extensions==2.1.3
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
kombu==4.2.1
mysqlclient==1.3.13
pycparser==2.19
pyOpenSSL==18.0.0
pytz==2018.7
requests==2.19.1
six==1.11.0
urllib3==1.23
uWSGI==2.0.17.1
vine==1.1.4
dj-static
#whitenoise==4.1.1
